test/LayoutTests/Panel/allInFile.html

<html>
<body>
<h1>Feature Test allInFile</h1>
</body>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/ChannelPlate/ChannelPlate.js"></script>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/ChannelPlate/RemoteMethodCall.js"></script>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/AsyncMachine.js"></script>
<script src="../../../DebugLogger.js"></script>
<script src="../PatientSelector.js"></script>
<script src="../panel-test.js"></script>

<script>

window.addEventListener('load', function() {
  testScript();
});

function test_lastChange() {
  console.log("------------test begins------------");
  AsyncMachine.pushOp(evaluateInPage, 'window.location="http://localhost:8686/test/qpd.html"');
  AsyncMachine.pushOp(openQuerypointPanel);
  AsyncMachine.pushOp(openSourceFile, 'demo.js');
  
  var objPropSelector = [
    {type: 'span.cm-string', text:'\"I\'ve been clicked \"'},
    {type: 'span.cm-property', text: 'prop'}
    ];
  AsyncMachine.pushOp(selectTokenInSource, objPropSelector);
  AsyncMachine.pushOp(verifyTokenView, 'obj.prop');
  AsyncMachine.pushOp(clickQPOperation, 'All in File');
  AsyncMachine.pushOp(whenSelectorAll, '.currentLoadNumber', '1');
  AsyncMachine.pushOp(evaluateInPage, 'document.querySelector("#myButton").dispatchEvent(new MouseEvent("click"))');
  AsyncMachine.pushOp(whenSelectorAll, 'td.traceValue', '0');
  AsyncMachine.pushOp(extractText, 'td.traceValue');

  // Test link to line traced
  AsyncMachine.pushOp(extractFromSelection, 'span.command', 'outerHTML', function(outerHTML){
    var m = /title="([^"]*)"/.exec(outerHTML);
    return m[1];
  });
  AsyncMachine.pushOp(clickSelector, 'span.command', '');
  AsyncMachine.pushOp(extractText, '.qp-highlight');

  AsyncMachine.runOp();
}

var testScript = test_lastChange;
</script>
</html>
