test/LayoutTests/Panel/allInFile-expected.txt

http://localhost:8686/test/qpd.html
prop in   var message = "I've been clicked " + obj.prop + " times";
true
0
All in File found in http://localhost:8686/test/demo.js
obj|.|prop
